ThanatosDi/requirements.txt
astroid==2.2.4
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
idna==2.8
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.6.1
pycryptodome==3.7.3
pylint==2.3.1
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
pycryptodome==3.7.3
